requirements.txt
 requests
numpy
librosa
aubio
praat-parselmouth
asyncpg
openai>=1.0.0
fastapi
hypercorn
python-socketio[asgi]
starlette
jinja2
python-multipart
slowapi
python-dotenv
